Move SystemTempl into NonlinearSystem
Move SystemTempl into NonlinearSystem to make AssemblySystem independent.

Any functions which depends on SysBase and SystemTempl, are moved back to
NonlinearSystem.

Refs #7398 #7654
